refactor: rename modal module to dialog
